+++ RayTracer/src/Geometry/Line.h
@@ -18,7 +18,7 @@ public:
 	bool intersect(Ray& ray, HitInfo& hit, unsigned int i) const;
 
 	/// return the number of primitives in the geometry
-	virtual size_t num_primitives() const;
+	virtual unsigned int num_primitives() const;
 	/// the AABB for the whole geometry
 	virtual AABB get_bbox() const;
 	/// the AABB for the primitive at index i
